--- src/main/java/com/study/sessionboard/infrastructure/EventPostRepository.java
@@ -1,8 +1,0 @@
-package com.study.sessionboard.infrastructure;
-
-import com.study.sessionboard.domain.event.EventPost;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface EventPostRepository extends JpaRepository<EventPost, Long> {}
