+++ src/main/java/com/study/sessionboard/infrastructure/event/EventPostImageRepository.java
@@ -20,4 +20,6 @@ public interface EventPostImageRepository extends JpaRepository<EventPostImage, 
   List<EventPostImage> findFirstImagesByPostIdIn(@Param("postIds") List<Long> postIds);
 
   void deleteByPostId(Long postId);
+
+  List<EventPostImage> findAllByPostIdOrderByOrderAsc(Long postId);
 }
